# backend_fashion_app/internal/common/httpx/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.AddToCartRequest:
    properties:
      product_variant_id:
        minimum: 1
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_variant_id
    - quantity
    type: object
...
  api.CartItemResponse:
    properties:
      created_at:
        type: string
      current_price:
        description: Giá hiện tại
        type: number
      id:
        description: CartItem ID
        type: integer
      price_changed:
        description: Có thay đổi giá không?
        type: boolean
      price_snapshot:
        description: Giá lúc thêm vào giỏ
        type: number
      product:
        $ref: '#/definitions/api.ProductInfo'
      product_variant_id:
        type: integer
      quantity:
        type: integer
      stock_issue:
        description: Số lượng trong giỏ > tồn kho?
        type: boolean
      subtotal:
        description: Tổng tiền = quantity * current_price
        type: number
    type: object
  api.CartResponse:
    properties:
      can_checkout:
        description: Có thể thanh toán không?
        type: boolean
      created_at:
        type: string
      has_issues:
        description: Có vấn đề gì không (giá thay đổi, hết hàng...)
        type: boolean
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/api.CartItemResponse'
        type: array
      total_amount:
        description: Tổng tiền
        type: number
      total_items:
        description: Tổng số sản phẩm (cộng dồn quantity)
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  api.CartSummaryResponse:
    properties:
      can_checkout:
        type: boolean
      issues:
        description: Danh sách các vấn đề (nếu có)
        items:
          type: string
        type: array
      items:
        items:
          $ref: '#/definitions/api.CartItemResponse'
        type: array
      total_amount:
        type: number
      total_items:
        type: integer
    type: object
...
  api.ChangeCartItemVariantRequest:
    properties:
      color_hex:
        type: string
      size_code:
        type: string
    type: object
...
  api.ProductInfo:
    properties:
      color:
        type: string
      current_price:
        description: Variant Price
        type: number
      image_url:
        description: Can be from Product or Variant
        type: string
      name:
        description: Product Name
        type: string
      product_id:
        type: integer
      size:
        type: string
      sku:
        description: Variant SKU
        type: string
      stock:
        description: Variant Stock
        type: integer
      variant_id:
        type: integer
    type: object
...
  api.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  resp.Envelope:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Auth (local, Google) + Profile (customer/shop) + Catalog
  title: MyFashion API
  version: "1.0"
paths:
...
  /cart:
    delete:
      description: Xóa toàn bộ sản phẩm trong giỏ hàng
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Clear Cart
      tags:
      - Cart
    get:
      description: Lấy thông tin giỏ hàng của người dùng hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Get Cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Thêm một sản phẩm vào giỏ hàng
      parameters:
      - description: Thông tin sản phẩm cần thêm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AddToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.CartItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Add Item to Cart
      tags:
      - Cart
  /cart/items/{id}:
    delete:
      description: Xóa một sản phẩm khỏi giỏ hàng
      parameters:
      - description: ID của Cart Item
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Remove Cart Item
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Cập nhật số lượng của một sản phẩm trong giỏ hàng
      parameters:
      - description: ID của Cart Item
        in: path
        name: id
        required: true
        type: integer
      - description: Số lượng mới
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.CartItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Update Cart Item
      tags:
      - Cart
  /cart/items/{id}/variant:
    put:
      consumes:
      - application/json
      description: Đổi màu/size cho một sản phẩm trong giỏ; sẽ gộp số lượng nếu cart
        đã có item cùng biến thể mới
      parameters:
      - description: ID của Cart Item
        in: path
        name: id
        required: true
        type: integer
      - description: Màu/Size mới
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ChangeCartItemVariantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.CartItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Change Cart Item Variant
      tags:
      - Cart
  /cart/summary:
    get:
      description: Lấy tóm tắt giỏ hàng để chuẩn bị thanh toán
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.CartSummaryResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Get Cart Summary
      tags:
      - Cart
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"